Do not run sdv vehicle data integration tests with binary PDK
As some of the dependencies of the integration tests are not needed
for the final product, these are also not collected for the binary
PDK. Therefore, compilation of the
"har_sdv_service_bundles_integration_tests" Soong module fails for
the binary PDK, to fix this problem, we are conditionally excluding
this module when we are building the binary PDK.

We also need to disable
“com.sdv.google.display_safety.services_bundle.apex” on the binary
PDK as it also depends on a dependency that is undefined in the
binary PDK (same reason as above). When building the
com.sdv.google.display_safety.services_bundle.apex, we run into the
case that it needs a dependency that we do not collect with the
binary PDk, as those are not (yet?) used when building the SDV
system.

The third case is the vehicle_data_publishers, which we
conditionally enable and add the dependency that is not collected
with the binary PDK. This is done via the soong_config_module_type,
which enables us to combine both steps into one and only include a
“defaults” definition.

Bug: 372173243
Test: apply binary PDK and run m droid dist
Flag: EXEMPT This product does not use flags
Change-Id: I32f21fa3bfe83bc6eb75d865fac96169773b3daf