Replace a deprecated method to find and connect RPC server
SdvGatewayClient.connectToRpcServerByName() is a method that identifies
an existing RPC server by given names and then returns a communication
channel to use it. Recently, its initial implementation got deprecated
and a new version of it is recommended to be used.

Also, with this new method, TLS configurations become transparent and
therefore this change consolidates methods to obtain a communication
channel to the remote RPC server.

Bug: 389375652
Test: Manually run SdvCameraPreviewApp on sdv_ivi_arm64_ds target and
      confirm it is able to send heartbeat messages.
Flag: EXEMPT bug fix
Change-Id: I57bea3f74c8df00ff7c7331f2ca149acf10ca89a