Add HAR update APEX
Create new HAR package that can be used as an update APEX.
The contents is still the same, except the version code is bumped.
Init.rc files were updated to make them updatable and the update APEX
has an override for the HAR service.

Bug: 349146763
Test: Not yet easily testable, but it builds and will be part of "dist".
Flag: EXEMPT Build file change to create an APEX update package
Change-Id: Ifffabee8bfd43ca07cd590e030b17a7b133b4bea